Fixes : OpenAPI spec & Test Script Changes (#460)
* feat(gamification): Initial commit of Badge/Achievement system.

Added core logic for awarding achievements and tracking metrics.

* Change: Switched to transformResponse and renamed services to lowerCamelCase.

Remove InstanceToPlain and use transformResponse Flag, change Services to follow lowerCamelCase.

* feat(Gamification): Added Lazy Init via Get or Create pattern.

Whenever a new metric is added and not already included in a user's metrics, it is automatically associated with that user via an upsert operation. If no user metrics exist at all, all available metrics are assigned to the user by default.

* OpenAPI spec & Test Script Changes